@@ -74,14 +74,15 @@ func (r ImageRequest) MarshalJSON() ([]byte, error) {
 		return nil, err
 	}
 
+	// 不能合并ExtraFields！！！！！！！！
 	// 合并 ExtraFields
-	for k, v := range r.Extra {
-		if _, exists := baseMap[k]; !exists {
-			baseMap[k] = v
-		}
-	}
+	//for k, v := range r.Extra {
+	//	if _, exists := baseMap[k]; !exists {
+	//		baseMap[k] = v
+	//	}
+	//}
 
-	return json.Marshal(baseMap)
+	return common.Marshal(baseMap)
 }
 
 func GetJSONFieldNames(t reflect.Type) map[string]struct{} {
