@@ -56,10 +56,10 @@ const MjLogsFilters = ({
               showClear
               pure
               size='small'
-              presets={DATE_RANGE_PRESETS.map(preset => ({
+              presets={DATE_RANGE_PRESETS.map((preset) => ({
                 text: t(preset.text),
                 start: preset.start(),
-                end: preset.end()
+                end: preset.end(),
               }))}
             />
           </div>
